OC-12637: Added super class ScheduleJobServlet to reduce redundancy and removed exit button from viewJobs
