+++ src/main/java/org/akaza/openclinica/control/admin/ViewJobServlet.java
@@ -1,64 +1,28 @@
 package org.akaza.openclinica.control.admin;
 
-import java.util.*;
-
 import core.org.akaza.openclinica.bean.admin.TriggerBean;
 import core.org.akaza.openclinica.bean.extract.DatasetBean;
-import core.org.akaza.openclinica.dao.core.CoreResources;
-import core.org.akaza.openclinica.dao.hibernate.StudyDao;
-import core.org.akaza.openclinica.domain.datamap.Study;
-import core.org.akaza.openclinica.job.AutowiringSpringBeanJobFactory;
-import core.org.akaza.openclinica.job.JobExecutionExceptionListener;
-import core.org.akaza.openclinica.job.JobTriggerListener;
-import core.org.akaza.openclinica.job.OpenClinicaSchedulerFactoryBean;
-import org.akaza.openclinica.control.SpringServletAccess;
-import org.akaza.openclinica.control.core.SecureController;
-import org.akaza.openclinica.control.form.FormProcessor;
 import core.org.akaza.openclinica.dao.extract.DatasetDAO;
+import core.org.akaza.openclinica.domain.datamap.Study;
 import core.org.akaza.openclinica.service.extract.XsltTriggerService;
-import org.akaza.openclinica.view.Page;
-import core.org.akaza.openclinica.web.InsufficientPermissionException;
 import core.org.akaza.openclinica.web.bean.EntityBeanTable;
 import core.org.akaza.openclinica.web.bean.TriggerRow;
 import core.org.akaza.openclinica.web.job.ExampleSpringJob;
-import org.apache.commons.lang.StringUtils;
+import org.akaza.openclinica.control.form.FormProcessor;
+import org.akaza.openclinica.view.Page;
 import org.quartz.*;
-import org.quartz.impl.StdScheduler;
 import org.quartz.impl.matchers.GroupMatcher;
-import org.springframework.beans.BeansException;
-import org.springframework.beans.factory.NoSuchBeanDefinitionException;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
 import org.springframework.context.ApplicationContext;
-import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.transaction.PlatformTransactionManager;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.sql.DataSource;
-
-import static core.org.akaza.openclinica.dao.hibernate.multitenant.CurrentTenantIdentifierResolverImpl.CURRENT_TENANT_ID;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.Set;
 
 /**
  * @author thickerson purpose: to generate the list of jobs and allow us to view them
  */
-public class ViewJobServlet extends SecureController {
-
-    @Override
-    protected void mayProceed() throws InsufficientPermissionException {
-        if (ub.isSysAdmin() || ub.isTechAdmin()) {
-            return;
-        }
-        addPageMessage(respage.getString("no_have_correct_privilege_current_study") + respage.getString("change_study_contact_sysadmin"));
-        throw new InsufficientPermissionException(Page.MENU_SERVLET, resexception.getString("not_allowed_access_extract_data_servlet"), "1");// TODO
-        // above copied from create dataset servlet, needs to be changed to
-        // allow only admin-level users
-    }
-
-    private StdScheduler getScheduler() {
-        scheduler = this.scheduler != null ? scheduler : (StdScheduler) SpringServletAccess.getApplicationContext(context).getBean(SCHEDULER);
-        return scheduler;
-    }
-
+public class ViewJobServlet extends ScheduleJobServlet {
 
     @Override
     protected void processRequest() throws Exception {
@@ -70,13 +34,12 @@ public class ViewJobServlet extends SecureController {
         // First we must get a reference to a scheduler
         ApplicationContext context = null;
         scheduler = getScheduler();
-        Scheduler jobScheduler;
         try {
             context = (ApplicationContext) scheduler.getContext().get("applicationContext");
         } catch (SchedulerException e) {
             logger.error("Error in receiving application context: ", e);
         }
-        jobScheduler = getSchemaScheduler(request, context, scheduler);
+        Scheduler jobScheduler = getSchemaScheduler(request, context, scheduler);
         XsltTriggerService xsltTriggerSrvc = new XsltTriggerService();
 
         Set<TriggerKey> triggerKeySet = jobScheduler.getTriggerKeys(GroupMatcher.triggerGroupEquals(xsltTriggerSrvc.getTriggerGroupNameForExportJobs()));
